// Game.Engine/wwwroot/src/network/game_generated.ts
export namespace Game.Engine.Networking.FlatBuffers {
  export enum AllMessages {
    NONE = 0,
    NetWorldView = 1,
    NetSpawn = 2,
    NetEvent = 3,
    NetControlInput = 4,
    NetPing = 5,
    NetLeaderboard = 6,
    NetExit = 7,
    NetAuthenticate = 8,
  }
}
// ...
  export class NetSpawn {
    bb: flatbuffers.ByteBuffer | null = null;

    bb_pos: number = 0;
    /**
     * @param number i
     * @param flatbuffers.ByteBuffer bb
     * @returns NetSpawn
     */
    __init(i: number, bb: flatbuffers.ByteBuffer): NetSpawn {
      this.bb_pos = i;
      this.bb = bb;
      return this;
    }

    /**
     * @param flatbuffers.ByteBuffer bb
     * @param NetSpawn= obj
     * @returns NetSpawn
     */
    static getRootAsNetSpawn(
      bb: flatbuffers.ByteBuffer,
      obj?: NetSpawn,
    ): NetSpawn {
      return (obj || new NetSpawn()).__init(
        bb.readInt32(bb.position()) + bb.position(),
        bb,
      );
    }

    /**
     * @param flatbuffers.Encoding= optionalEncoding
     * @returns string|Uint8Array|null
     */
    name(): string | null;
    name(optionalEncoding: flatbuffers.Encoding): string | Uint8Array | null;
    name(optionalEncoding?: any): string | Uint8Array | null {
      var offset = this.bb!.__offset(this.bb_pos, 4);
      return offset
        ? this.bb!.__string(this.bb_pos + offset, optionalEncoding)
        : null;
    }

    /**
     * @param flatbuffers.Encoding= optionalEncoding
     * @returns string|Uint8Array|null
     */
    ship(): string | null;
    ship(optionalEncoding: flatbuffers.Encoding): string | Uint8Array | null;
    ship(optionalEncoding?: any): string | Uint8Array | null {
      var offset = this.bb!.__offset(this.bb_pos, 6);
      return offset
        ? this.bb!.__string(this.bb_pos + offset, optionalEncoding)
        : null;
    }

    /**
     * @param flatbuffers.Encoding= optionalEncoding
     * @returns string|Uint8Array|null
     */
    color(): string | null;
    color(optionalEncoding: flatbuffers.Encoding): string | Uint8Array | null;
    color(optionalEncoding?: any): string | Uint8Array | null {
      var offset = this.bb!.__offset(this.bb_pos, 8);
      return offset
        ? this.bb!.__string(this.bb_pos + offset, optionalEncoding)
        : null;
    }

    /**
     * @param flatbuffers.Encoding= optionalEncoding
     * @returns string|Uint8Array|null
     */
    token(): string | null;
    token(optionalEncoding: flatbuffers.Encoding): string | Uint8Array | null;
    token(optionalEncoding?: any): string | Uint8Array | null {
      var offset = this.bb!.__offset(this.bb_pos, 10);
      return offset
        ? this.bb!.__string(this.bb_pos + offset, optionalEncoding)
        : null;
    }

    /**
     * @param flatbuffers.Builder builder
     */
    static startNetSpawn(builder: flatbuffers.Builder) {
      builder.startObject(4);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param flatbuffers.Offset nameOffset
     */
    static addName(
      builder: flatbuffers.Builder,
      nameOffset: flatbuffers.Offset,
    ) {
      builder.addFieldOffset(0, nameOffset, 0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param flatbuffers.Offset shipOffset
     */
    static addShip(
      builder: flatbuffers.Builder,
      shipOffset: flatbuffers.Offset,
    ) {
      builder.addFieldOffset(1, shipOffset, 0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param flatbuffers.Offset colorOffset
     */
    static addColor(
      builder: flatbuffers.Builder,
      colorOffset: flatbuffers.Offset,
    ) {
      builder.addFieldOffset(2, colorOffset, 0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param flatbuffers.Offset tokenOffset
     */
    static addToken(
      builder: flatbuffers.Builder,
      tokenOffset: flatbuffers.Offset,
    ) {
      builder.addFieldOffset(3, tokenOffset, 0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @returns flatbuffers.Offset
     */
    static endNetSpawn(builder: flatbuffers.Builder): flatbuffers.Offset {
      var offset = builder.endObject();
      return offset;
    }
  }
// ...
  export class NetControlInput {
    bb: flatbuffers.ByteBuffer | null = null;

    bb_pos: number = 0;
    /**
     * @param number i
     * @param flatbuffers.ByteBuffer bb
     * @returns NetControlInput
     */
    __init(i: number, bb: flatbuffers.ByteBuffer): NetControlInput {
      this.bb_pos = i;
      this.bb = bb;
      return this;
    }

    /**
     * @param flatbuffers.ByteBuffer bb
     * @param NetControlInput= obj
     * @returns NetControlInput
     */
    static getRootAsNetControlInput(
      bb: flatbuffers.ByteBuffer,
      obj?: NetControlInput,
    ): NetControlInput {
      return (obj || new NetControlInput()).__init(
        bb.readInt32(bb.position()) + bb.position(),
        bb,
      );
    }

    /**
     * @returns number
     */
    angle(): number {
      var offset = this.bb!.__offset(this.bb_pos, 4);
      return offset ? this.bb!.readFloat32(this.bb_pos + offset) : 0.0;
    }

    /**
     * @returns number
     */
    x(): number {
      var offset = this.bb!.__offset(this.bb_pos, 6);
      return offset ? this.bb!.readFloat32(this.bb_pos + offset) : 0.0;
    }

    /**
     * @returns number
     */
    y(): number {
      var offset = this.bb!.__offset(this.bb_pos, 8);
      return offset ? this.bb!.readFloat32(this.bb_pos + offset) : 0.0;
    }

    /**
     * @returns boolean
     */
    boost(): boolean {
      var offset = this.bb!.__offset(this.bb_pos, 10);
      return offset ? !!this.bb!.readInt8(this.bb_pos + offset) : false;
    }

    /**
     * @returns boolean
     */
    shoot(): boolean {
      var offset = this.bb!.__offset(this.bb_pos, 12);
      return offset ? !!this.bb!.readInt8(this.bb_pos + offset) : false;
    }

    /**
     * @param flatbuffers.Encoding= optionalEncoding
     * @returns string|Uint8Array|null
     */
    spectateControl(): string | null;
    spectateControl(
      optionalEncoding: flatbuffers.Encoding,
    ): string | Uint8Array | null;
    spectateControl(optionalEncoding?: any): string | Uint8Array | null {
      var offset = this.bb!.__offset(this.bb_pos, 14);
      return offset
        ? this.bb!.__string(this.bb_pos + offset, optionalEncoding)
        : null;
    }

    /**
     * @param flatbuffers.Encoding= optionalEncoding
     * @returns string|Uint8Array|null
     */
    customData(): string | null;
    customData(
      optionalEncoding: flatbuffers.Encoding,
    ): string | Uint8Array | null;
    customData(optionalEncoding?: any): string | Uint8Array | null {
      var offset = this.bb!.__offset(this.bb_pos, 16);
      return offset
        ? this.bb!.__string(this.bb_pos + offset, optionalEncoding)
        : null;
    }

    /**
     * @param flatbuffers.Builder builder
     */
    static startNetControlInput(builder: flatbuffers.Builder) {
      builder.startObject(7);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param number angle
     */
    static addAngle(builder: flatbuffers.Builder, angle: number) {
      builder.addFieldFloat32(0, angle, 0.0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param number x
     */
    static addX(builder: flatbuffers.Builder, x: number) {
      builder.addFieldFloat32(1, x, 0.0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param number y
     */
    static addY(builder: flatbuffers.Builder, y: number) {
      builder.addFieldFloat32(2, y, 0.0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param boolean boost
     */
    static addBoost(builder: flatbuffers.Builder, boost: boolean) {
      builder.addFieldInt8(3, +boost, +false);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param boolean shoot
     */
    static addShoot(builder: flatbuffers.Builder, shoot: boolean) {
      builder.addFieldInt8(4, +shoot, +false);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param flatbuffers.Offset spectateControlOffset
     */
    static addSpectateControl(
      builder: flatbuffers.Builder,
      spectateControlOffset: flatbuffers.Offset,
    ) {
      builder.addFieldOffset(5, spectateControlOffset, 0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param flatbuffers.Offset customDataOffset
     */
    static addCustomData(
      builder: flatbuffers.Builder,
      customDataOffset: flatbuffers.Offset,
    ) {
      builder.addFieldOffset(6, customDataOffset, 0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @returns flatbuffers.Offset
     */
    static endNetControlInput(
      builder: flatbuffers.Builder,
    ): flatbuffers.Offset {
      var offset = builder.endObject();
      return offset;
    }
  }
// ...
  export class NetEvent {
    bb: flatbuffers.ByteBuffer | null = null;

    bb_pos: number = 0;
    /**
     * @param number i
     * @param flatbuffers.ByteBuffer bb
     * @returns NetEvent
     */
    __init(i: number, bb: flatbuffers.ByteBuffer): NetEvent {
      this.bb_pos = i;
      this.bb = bb;
      return this;
    }

    /**
     * @param flatbuffers.ByteBuffer bb
     * @param NetEvent= obj
     * @returns NetEvent
     */
    static getRootAsNetEvent(
      bb: flatbuffers.ByteBuffer,
      obj?: NetEvent,
    ): NetEvent {
      return (obj || new NetEvent()).__init(
        bb.readInt32(bb.position()) + bb.position(),
        bb,
      );
    }

    /**
     * @param flatbuffers.Encoding= optionalEncoding
     * @returns string|Uint8Array|null
     */
    type(): string | null;
    type(optionalEncoding: flatbuffers.Encoding): string | Uint8Array | null;
    type(optionalEncoding?: any): string | Uint8Array | null {
      var offset = this.bb!.__offset(this.bb_pos, 4);
      return offset
        ? this.bb!.__string(this.bb_pos + offset, optionalEncoding)
        : null;
    }

    /**
     * @param flatbuffers.Encoding= optionalEncoding
     * @returns string|Uint8Array|null
     */
    data(): string | null;
    data(optionalEncoding: flatbuffers.Encoding): string | Uint8Array | null;
    data(optionalEncoding?: any): string | Uint8Array | null {
      var offset = this.bb!.__offset(this.bb_pos, 6);
      return offset
        ? this.bb!.__string(this.bb_pos + offset, optionalEncoding)
        : null;
    }

    /**
     * @param flatbuffers.Builder builder
     */
    static startNetEvent(builder: flatbuffers.Builder) {
      builder.startObject(2);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param flatbuffers.Offset typeOffset
     */
    static addType(
      builder: flatbuffers.Builder,
      typeOffset: flatbuffers.Offset,
    ) {
      builder.addFieldOffset(0, typeOffset, 0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param flatbuffers.Offset dataOffset
     */
    static addData(
      builder: flatbuffers.Builder,
      dataOffset: flatbuffers.Offset,
    ) {
      builder.addFieldOffset(1, dataOffset, 0);
    }

    /**
     * @param flatbuffers.Builder builder
     * @returns flatbuffers.Offset
     */
    static endNetEvent(builder: flatbuffers.Builder): flatbuffers.Offset {
      var offset = builder.endObject();
      return offset;
    }
  }